Modified the Update StatisticalSourceListView to show all stat methods if the the logged in user is in the nemi_admin group.
Because the user is not necessarily the person who inserted the method now, overrode the  get_context_data method in UpdateStatisticalSourceView and added the context key 'insert_user' and assigned it the object's insert_person. Then modified the base_edit_stats_method to check for this context variable and if present use it rather than the user context variable.