Added code to test_runner to include auto managed models (i.e. M2M) so the models which are changed to managed from unmanaged during test execution.
Added a new fixture source_citation_ref which contains source citation ref data to be used for the statistical method page tests.

Created Jenkins specific settings in settings.py. The Jenkins specific settings are only set if JENKINS_URL is defined in the environment.

Created a Jenkins specific test runner which is the same as test_runner with the exception that it inherits from django_jenkins.runner.CITestSuiteRunner