Eliminated the KeywordSearchForm and implemented the form directly in the template file.
Modified the keyword search view to return an error context variable when the keyword field is blank. The template uses this to display an error message.

Put a keyword search field in the upper right of every search page. Put this in base.html inside a block so that the display of the keyword search section can be specified on each individual page within the project.

Updated the keyword search functional tests to add remove checks for the form and add a test for the new error context variable.