Refactored the site navigation. Now use a navigation bar near the top of the page which contains nav buttons. Some of the buttons pop down menus for additional choices. I am using the jquery.dropdownPlain plug in to implement the navigation menus.
Added a h1 heading bar above the navigation which contains the name of the NEMI page, so that the user can tell at a glance what page they are one. 

The keyword search appear to the far left of the navigation menu, visually both are in a colored bar below the h1 heading bar.

Added a very preliminary Home page.

Cleaned up template comments.