Refactored to have different results views for different classes of results. We now have distinct results pages for analyte searches, for statistical method searches, and for all other searches.
Implementation of the view changed to use extend ListView to ResultsView which includes the common queries for each result type, such as category, subcategory, and the Limit by filters. The ResultsView was extended to implement the specific views, defining the base query set to use and then overriding and extending the get_queryset method to add the specific filters for that kind of search.