Added a Compare Selected Results button to the results pages (except for the browse and keyword).
Modified the way the tablesorter plugin is used to make use of the meta-data and class specifications for plugin features such as what columns to sort and what parsers to use. Specified the Method Number column to use text as it was using digit as the default. This makes the sorting make more sense.