Added a new app, newsfeed. The app contains a single model, NewsItem, containing a headline and a created and modified date fields.
The app also contains an admin so that newsitems can be managed from the admin interface. The headline field uses the tinymce javascript WYSIWYG editor. Added the django-tinymce app to the list of requirements. I am using its built in TinyMCE widget for the headline editing field in the admin.

For this project, I want the latest news on all pages, so I added a context_processor to do return the last two years of news items. This seems like a reasonable limit. Added unit tests for this new context processor and added it and the tests to settings.py.