Added FieldHelpMixin to the MethodResultsView and AnalyteResultsView to provide popup help for column descriptions.
Moved the help dialog javascript to it's own file since these popup now appear in several pages.

Modified the html/css so that when pop up help is available for results columns, the help button appears below the column header label. Also had to modify the javascript so that when the user clicks on the pop up help button, the help appears but no sorting happens. I use to be able to do this with a click handler, but the tablesorter plugin now uses buttondown and buttonup handlers so I had to change my javascript to stop propogation of these events.