Created a new general purpose view, PdfView, in the common app which contains attributes for mimetype, pdf, and filename. If these are not specified when extending the class, then the user should override get_response_info to set these attributes. The view responses to get requests with a response containing the pdf file with mimetype. Wrote tests for this view.
Extended PdfView for MethodPdfView and create RevisionPdfView to provide pdf download capability on the method summary page and revision tab. 

Added latest_revision and revisions to the object returned in the MethodSummaryView. The latest_revision will always be the first one listed. We are using
descending order on the revisions to display the rest of them, filtering out the latest.