Added the flatpages app to the project. It's admin is in the common project and urls in nemi_project.urls.
The Tinymce editors for the places where it is used is now customized. I changed the defaults settings to use the advanced theme and this is what the flatpages app uses to edit content. The Help Content uses a subset of the buttons and the news feed app uses just a few to do simple formatting but added subscripts and superscripts.