Added the django.js app so that I can run jasmine tests from the browser and from the jenkins test suite.
Added Leaflet to the javascript library for this application.

Added javascript code which creates a WQP siteLayer with sensible
defaults. All defaults can be overridden. I have also provided a
function create a default map along with the layer. I see this as usable
in this application but generally I see users of this code more likely
to use the siteLayer function and then to put that layer into there
custom map.